Modified Periodic Tasks to Instant Execution Commented out RestrictedPython code execution Fix PrimitivePointer error in object search service
