+++ packages/grid/backend/grid/worker.py
@@ -1,12 +1,7 @@
 # stdlib
-from io import StringIO
-import sys
 from typing import Any
-from typing import Dict
 
 # third party
-from RestrictedPython import compile_restricted
-from RestrictedPython.Guards import safe_builtins
 from celery.utils.log import get_task_logger
 
 # syft absolute
@@ -28,7 +23,6 @@ from syft.core.store.storeable_object import StorableObject
 from grid.core.celery_app import celery_app
 from grid.core.config import settings  # noqa: F401
 from grid.core.node import node
-from grid.periodic_tasks import check_tasks_to_be_executed
 from grid.periodic_tasks import cleanup_incomplete_uploads_from_blob_store
 
 logger = get_task_logger(__name__)
@@ -48,110 +42,6 @@ def msg_without_reply(self, obj_msg: Any) -> None:  # type: ignore
         raise Exception(
             f"This worker can only handle SignedImmediateSyftMessageWithoutReply. {obj_msg}"
         )
-
-
-stdout_ = sys.stdout
-stderr_ = sys.stderr
-
-
-restricted_globals = dict(__builtins__=safe_builtins)
-
-
-@celery_app.task
-def execute_task(
-    task_uid: str, code: str, inputs: Dict[str, str], outputs: Dict[str, str]
-) -> None:
-    global stdout_
-    global stderr_
-    try:
-        node.tasks.update(
-            search_params={"uid": task_uid},
-            updated_args={"execution": {"status": "executing"}},
-        )
-
-        # Check overlap between inputs and vars
-        global_input_inter = set(globals().keys()).intersection(set(inputs.keys()))
-        local_input_inter = set(vars().keys()).intersection(set(inputs.keys()))
-
-        # If there's some intersection between global variables and input
-        if global_input_inter or local_input_inter:
-            stderr_message = " You can't use variable name: "
-            stderr_message += ",".join(list(global_input_inter))
-            stderr_message += ",".join(list(local_input_inter))
-
-            node.tasks.update(
-                search_params={"uid": task_uid},
-                updated_args={
-                    "execution": {"status": "failed", "stderr": stderr_message}
-                },
-            )
-            return None
-
-        local_vars = {}
-        for key, value in inputs.items():
-            local_vars[key] = node.store.get(value, proxy_only=True).data
-
-        # create file-like string to capture ouputs
-        codeOut = StringIO()
-        codeErr = StringIO()
-
-        sys.stdout = codeOut
-        sys.stderr = codeErr
-
-        old_env = set(vars().keys())
-
-        byte_code = compile_restricted(code, "<string>", "exec")
-        exec(byte_code, restricted_globals)
-
-        method_name = list(
-            set([key for key in vars().keys() if key != "old_env"]) - old_env
-        )[0]
-        outputs = vars()[method_name](**local_vars)
-
-        # restore stdout and stderr
-        sys.stdout = stdout_
-        sys.stderr = stderr_
-
-        logger.info(outputs)
-
-        logger.info("Error: " + str(codeErr.getvalue()))
-        logger.info("Std ouputs: " + str(codeOut.getvalue()))
-
-        # _save_vars = {}
-        # for var in outputs.keys():
-        new_id = UID()
-        node.store.check_collision(new_id)
-
-        obj = StorableObject(
-            id=new_id,
-            data=outputs,
-            search_permissions={VERIFYALL: None},
-        )
-
-        obj.read_permissions = {
-            node.verify_key: node.id,
-        }
-
-        obj.write_permissions = {
-            node.verify_key: node.id,
-        }
-
-        node.store[new_id] = obj
-
-        node.tasks.update(
-            search_params={"uid": task_uid},
-            updated_args={
-                "execution": {"status": "done"},
-                "outputs": {"output": new_id.to_string()},
-            },
-        )
-    except Exception as e:
-        sys.stdout = stdout_
-        sys.stderr = stderr_
-        print("Task Failed with Exception", e)
-    finally:
-        sys.stdout = stdout_
-        sys.stderr = stderr_
 
 
 @celery_app.task
@@ -211,17 +101,6 @@ def add_cleanup_blob_store_periodic_task(sender, **kwargs) -> None:  # type: ign
         3600,  # Run every hour
         cleanup_incomplete_uploads_from_blob_store.s(),
         name="Clean incomplete uploads in Seaweed",
-        queue="main-queue",
-        options={"queue": "main-queue"},
-    )
-
-
-@celery_app.on_after_configure.connect
-def check_ready_tasks(sender, **kwargs) -> None:  # type: ignore
-    celery_app.add_periodic_task(
-        3,  # Run every hour
-        check_tasks_to_be_executed.s(),
-        name="Check tasks that are ready to be executed ...",
         queue="main-queue",
         options={"queue": "main-queue"},
     )
