+++ packages/syft/src/syft/core/node/common/node_service/object_search/obj_search_service.py
@@ -14,6 +14,7 @@ from nacl.signing import VerifyKey
 from typing_extensions import final
 
 # relative
+from ......lib.python.util import downcast
 from ......logger import error
 from ......util import obj2pointer_type
 from ......util import traceback_and_raise
@@ -120,11 +121,11 @@ class ImmediateObjectSearchService(ImmediateNodeServiceWithReply):
                         )
 
                     else:
-                        if hasattr(obj.data, "init_pointer"):
-                            ptr_constructor = obj.data.init_pointer  # type: ignore
+                        obj_data = downcast(obj.data)
+                        if hasattr(obj_data, "init_pointer"):
+                            ptr_constructor = obj_data.init_pointer  # type: ignore
                         else:
-                            ptr_constructor = obj2pointer_type(obj=obj.data)
-
+                            ptr_constructor = obj2pointer_type(obj=obj_data)
                         ptr = ptr_constructor(
                             client=node,
                             id_at_location=obj.id,
