+++ packages/syft/src/syft/core/node/common/node_service/task_submission/task_submission.py
@@ -1,5 +1,7 @@
 # stdlib
 from datetime import date
+from io import StringIO
+import sys
 from typing import Any
 from typing import Dict
 from typing import List
@@ -11,8 +13,12 @@ from nacl.signing import VerifyKey
 from typing_extensions import final
 
 # relative
+from ...... import logger
 from ......core.common.uid import UID
+from ......core.store.storeable_object import StorableObject
+from .....common.group import VERIFYALL
 from .....common.serde.serializable import serializable
+from ....abstract.node import AbstractNode
 from ....domain_interface import DomainInterface
 from ....domain_msg_registry import DomainMessageRegistry
 from ...node_table.task import NoSQLTask
@@ -26,6 +32,9 @@ from .enum import EXECUTION_STATUS
 from .enum import TASK_SERVICE_DEFAULT_MESSAGES
 from .enum import TASK_SERVICE_FIELDS
 from .enum import TASK_SERVICE_STATUS
+
+# from RestrictedPython.Guards import safe_builtins
+# from RestrictedPython import compile_restricted
 
 
 @serializable(recursive_serde=True)
@@ -212,9 +221,116 @@ class ReviewTask(SyftMessage, DomainMessageRegistry):
             search_params={TASK_SERVICE_FIELDS.UID.value: self.payload.task_uid},
             updated_args=update_values,
         )
+        task = node.tasks.find(search_params={"uid": self.payload.task_uid})
+        if not task:
+            raise Exception(f"The given task_id:{self.payload.task_uid} does not exist")
+        task = task[0]
+        execute_task(node, task.uid, task.code, task.inputs, task.outputs)
 
         return CreateTask.Reply()
 
     def get_permissions(self) -> List[Type[BasePermission]]:
         """Returns the list of permission classes."""
         return [UserIsOwner]
+
+
+stdout_ = sys.stdout
+stderr_ = sys.stderr
+
+
+def execute_task(
+    node: AbstractNode,
+    task_uid: str,
+    code: str,
+    inputs: Dict[str, str],
+    outputs: Dict[str, str],
+) -> None:
+    global stdout_
+    global stderr_
+    try:
+        logger.info(f"inital outputs: {outputs}")
+        node.tasks.update(
+            search_params={"uid": task_uid},
+            updated_args={"execution": {"status": "executing"}},
+        )
+
+        # Check overlap between inputs and vars
+        global_input_inter = set(globals().keys()).intersection(set(inputs.keys()))
+        local_input_inter = set(vars().keys()).intersection(set(inputs.keys()))
+
+        # If there's some intersection between global variables and input
+        if global_input_inter or local_input_inter:
+            stderr_message = " You can't use variable name: "
+            stderr_message += ",".join(list(global_input_inter))
+            stderr_message += ",".join(list(local_input_inter))
+
+            node.tasks.update(
+                search_params={"uid": task_uid},
+                updated_args={
+                    "execution": {"status": "failed", "stderr": stderr_message}
+                },
+            )
+            return None
+
+        local_vars = {}
+        for key, value in inputs.items():
+            local_vars[key] = node.store.get(value, proxy_only=True).data
+
+        # create file-like string to capture ouputs
+        codeOut = StringIO()
+        codeErr = StringIO()
+
+        sys.stdout = codeOut
+        sys.stderr = codeErr
+
+        locals().update(local_vars)
+        # byte_code = compile_restricted(code, "<string>", "exec")
+        # exec(byte_code, restricted_globals)
+        exec(code)
+
+        for output in outputs:
+            logger.info(f"variable: {output} result: {vars()[output]}")
+            outputs[output] = vars()[output]
+
+        # restore stdout and stderr
+        sys.stdout = stdout_
+        sys.stderr = stderr_
+
+        logger.info(outputs)
+
+        logger.info("Error: " + str(codeErr.getvalue()))
+        logger.info("Std ouputs: " + str(codeOut.getvalue()))
+
+        new_id = UID()
+        node.store.check_collision(new_id)
+
+        obj = StorableObject(
+            id=new_id,
+            data=outputs,
+            search_permissions={VERIFYALL: None},
+        )
+
+        obj.read_permissions = {
+            node.verify_key: node.id,
+        }
+
+        obj.write_permissions = {
+            node.verify_key: node.id,
+        }
+
+        node.store[new_id] = obj
+
+        node.tasks.update(
+            search_params={"uid": task_uid},
+            updated_args={
+                "execution": {"status": "done"},
+                "outputs": {"output": new_id.to_string()},
+            },
+        )
+    except Exception as e:
+        sys.stdout = stdout_
+        sys.stderr = stderr_
+        print("Task Failed with Exception", e)
+    finally:
+        sys.stdout = stdout_
+        sys.stderr = stderr_
